# Improve bulk series file writes.