--- a/pkg/rhh/rhh.go
+++ b/pkg/rhh/rhh.go
@@ -41,6 +41,8 @@ func (m *HashMap) Reset() {
 	m.n = 0
 }
 
+func (m *HashMap) LoadFactor() int { return m.loadFactor }
+
 func (m *HashMap) Get(key []byte) interface{} {
 	i := m.index(key)
 	if i == -1 {
@@ -53,7 +55,7 @@ func (m *HashMap) Put(key []byte, val interface{}) {
 	// Grow the map if we've run out of slots.
 	m.n++
 	if m.n > m.threshold {
-		m.grow()
+		m.Grow(m.capacity * 2)
 	}
 
 	// If the key was overwritten then decrement the size.
@@ -120,14 +122,20 @@ func (m *HashMap) alloc() {
 	m.mask = int64(m.capacity - 1)
 }
 
-// grow doubles the capacity and reinserts all existing hashes & elements.
-func (m *HashMap) grow() {
+// Grow increases the capacity and reinserts all existing hashes & elements.
+func (m *HashMap) Grow(sz int64) {
+	// Ensure new capacity is a power of two and greater than current capacity.
+	sz = pow2(sz)
+	if sz <= m.capacity {
+		return
+	}
+
 	// Copy old elements and hashes.
 	elems, hashes := m.elems, m.hashes
 	capacity := m.capacity
 
-	// Double capacity & reallocate.
-	m.capacity *= 2
+	// Increase capacity & reallocate.
+	m.capacity = sz
 	m.alloc()
 
 	// Copy old elements to new hash/elem list.