_Briefly describe your proposed changes:_

This pull request preallocates space in the RHH hashmap before we bulk import into a partition. Normally the hashmap would keep doubling in space but that can take many iterations for large imports.

Previous benchmark:

```
BenchmarkSeriesPartition_CreateSeriesListIfNotExists/1000000-4         	       1	3611973385 ns/op
```

New benchmark:

```
BenchmarkSeriesPartition_CreateSeriesListIfNotExists/1000000-4         	       1	2519296210 ns/op
```

There is an approximately 30% improvement per partition (`3.6s` to `2.5s`) when performing a large import. The preallocation also occurs in a staggered lock so that the continuous lock time is broken up.

_What was the problem?_

We wanted to improve bulk import performance.

_What was the solution?_

Preallocate hashmap space and remove duplicate effort caused by constantly resizing the hashmap.


  - [x] Rebased/mergeable
  - [x] Tests pass