@mark-rushakoff 2019-01-11
Closing since we've moved to the influxdb repository. If you need help transferring this PR, here's a simple way to do it:

1. Download the patch from GitHub, by adding .patch to the end of the pull request URL, e.g. wget -P /tmp https://github.com/influxdata/platform/pull/1599.patch to download 1599.patch to /tmp (so that you don't accidentally check in 1599.patch in your influxdb directory).
1. Check out the new influxdb branch you'll use for the PR, like any time you start new work.
1. Use git-am to apply the commits in the patch: git am /tmp/1599.patch. Then push the branch to GitHub and open a PR like you would normally.